% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{df_groups}
\alias{df_groups}
\title{df_groups}
\usage{
df_groups(df_participants_user, attributes_groups)
}
\arguments{
\item{df_participants_user}{\dots}

\item{attributes_groups}{\dots

internal function}
}
\description{
groups data frame
}
\keyword{internal}
