% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_contacts_panel.R
\name{mod_contacts_panel_ui}
\alias{mod_contacts_panel_ui}
\alias{mod_contacts_panel_server}
\title{mod_contacts_panel_ui and mod_contacts_panel_server}
\usage{
mod_contacts_panel_ui(id)

mod_contacts_panel_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
