% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_groups_panel.R
\name{mod_groups_panel_ui}
\alias{mod_groups_panel_ui}
\alias{mod_groups_panel_server}
\title{mod_groups_panel_ui and mod_groups_panel_server}
\usage{
mod_groups_panel_ui(id)

mod_groups_panel_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
