% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_init_sqlite.R
\name{mod_init_sqlite_ui}
\alias{mod_init_sqlite_ui}
\alias{mod_init_sqlite_server}
\title{mod_init_sqlite_ui and mod_init_sqlite_server}
\usage{
mod_init_sqlite_ui(id)

mod_init_sqlite_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
