% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(sqlite_base, cron_responses, title = "Phoning")
}
\arguments{
\item{sqlite_base}{SQLite base path}

\item{cron_responses}{CRON responses rda path}

\item{title}{Shiny application title}
}
\description{
Run the Shiny Application
}
