% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{df_participants_events}
\alias{df_participants_events}
\title{df_participants_events}
\usage{
df_participants_events(df)
}
\arguments{
\item{df}{\dots

internal function}
}
\description{
participants event summary data frame
}
\keyword{internal}
